# frontend/react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upstream API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tracking": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Tracking",
        "description": "List tracking entries for a colleague.\n\nArgs:\n    colleague_id: Optional colleague username to query.\n    current_user: Authenticated username.\n\nReturns:\n    list[dict]: Tracking entries.",
        "operationId": "get_tracking_tracking_get",
        "parameters": [
          {
            "name": "colleague_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colleague Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackingRecordPayload"
                  },
                  "title": "Response Get Tracking Tracking Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tracking"
        ],
        "summary": "Set Tracking",
        "description": "Create or update tracking status for a course.\n\nArgs:\n    payload: Tracking payload with course_id and status.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Tracking record.",
        "operationId": "set_tracking_tracking_post",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingRecordPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tracking/delete": {
      "post": {
        "tags": [
          "tracking"
        ],
        "summary": "Delete Tracking",
        "description": "Remove tracking status for a course.\n\nArgs:\n    payload: Delete payload with course_id.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Delete result.",
        "operationId": "delete_tracking_tracking_delete_post",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tracking/stats": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Stats",
        "description": "Return self-only statistics, or aggregate statistics for administrators.\n\nArgs:\n    colleague_id: Optional colleague username to query.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Stats payload.",
        "operationId": "get_stats_tracking_stats_get",
        "parameters": [
          {
            "name": "colleague_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Colleague Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Get Stats Tracking Stats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tracking/stats/users": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Stats By User",
        "description": "Return tracking statistics grouped by user for administrators.\n\nArgs:\n    current_user: Authenticated username.\n\nReturns:\n    list[dict]: Stats by user.",
        "operationId": "get_stats_by_user_tracking_stats_users_get",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackingStatsByUserItem"
                  },
                  "title": "Response Get Stats By User Tracking Stats Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tracking/recent": {
      "get": {
        "tags": [
          "tracking"
        ],
        "summary": "Get Recent Activity",
        "description": "Return recent tracking activity (admin only).\n\nArgs:\n    limit: Max number of records.\n    current_user: Authenticated username.\n\nReturns:\n    list[dict]: Recent activity records.",
        "operationId": "get_recent_activity_tracking_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackingRecordPayload"
                  },
                  "title": "Response Get Recent Activity Tracking Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrackingDeleteRequest": {
        "properties": {
          "course_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          }
        },
        "type": "object",
        "title": "TrackingDeleteRequest",
        "description": "Tracking delete request payload."
      },
      "TrackingDeleteResponse": {
        "properties": {
          "removed": {
            "type": "integer",
            "title": "Removed"
          }
        },
        "type": "object",
        "required": [
          "removed"
        ],
        "title": "TrackingDeleteResponse",
        "description": "Tracking delete response payload."
      },
      "TrackingRecordPayload": {
        "properties": {
          "colleague_id": {
            "type": "string",
            "title": "Colleague Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "colleague_id",
          "course_id",
          "status",
          "updated_at"
        ],
        "title": "TrackingRecordPayload",
        "description": "Tracking record response payload."
      },
      "TrackingStatsByUserItem": {
        "properties": {
          "colleague_id": {
            "type": "string",
            "title": "Colleague Id"
          },
          "interested": {
            "type": "integer",
            "title": "Interested"
          },
          "in_progress": {
            "type": "integer",
            "title": "In Progress"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "colleague_id",
          "interested",
          "in_progress",
          "completed"
        ],
        "title": "TrackingStatsByUserItem",
        "description": "Tracking stats-by-user item payload."
      },
      "TrackingUpsertRequest": {
        "properties": {
          "course_id": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "TrackingUpsertRequest",
        "description": "Tracking upsert request payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}